1545:renaming variables folder to constants
