+++ components/molecules/auth-button/AuthButton.constants.ts
@@ -1,3 +1,3 @@
-import { ACCESS_UPPER_UZ } from 'variables/common';
+import { ACCESS_UPPER_UZ } from 'constants/common';
 
 export const TRUSTED_ORIGINS = [ACCESS_UPPER_UZ];
