+++ constants/common.ts
