+++ frontends/landing/body/Body.tsx
@@ -11,11 +11,11 @@ import {
 } from 'components/molecules';
 import { Sidebar } from 'components/organisms';
 import { COMMON_SPACE_FROM_TOP } from 'components/wrappers';
+import { PopularLabels } from 'constants/labels';
 import { useUrlParams } from 'hooks';
 import { usePublishedArticlesList } from 'store/clients/published-article';
 import { getClassName } from 'utils';
 import { addAmazonBucketUrl } from 'utils/published-article';
-import { PopularLabels } from 'variables';
 
 import classes from './Body.module.scss';
