+++ hooks/useAppRouter/useAppRouter.ts
@@ -1,6 +1,6 @@
+import { WEB_APP_ROOT_DIR } from 'constants/common';
 import { NextRouter, useRouter } from 'next/router';
 import { useCallback, useMemo } from 'react';
-import { WEB_APP_ROOT_DIR } from 'variables';
 
 export const useAppRouter = (): NextRouter => {
   const router = useRouter();
