+++ types/images.ts
@@ -1,3 +1,3 @@
-import { IMAGE_TYPES } from 'variables';
+import { IMAGE_TYPES } from 'constants/images';
 
 export type TImage = keyof typeof IMAGE_TYPES;
