--- variables/index.ts
@@ -1,11 +1,0 @@
-export * from './notification';
-export * from './common';
-export * from './Blog.constants';
-export * from './auth';
-export * from './images';
-export * from './article';
-export * from './editor';
-export * from './published-article';
-export * from './theme';
-export * from './emoji-select';
-export * from './labels';
