+++ vitest.config.ts
@@ -32,8 +32,7 @@ export default defineConfig({
       hooks: path.resolve(__dirname, './hooks'),
       utils: path.resolve(__dirname, './utils'),
       store: path.resolve(__dirname, './store'),
-      variables: path.resolve(__dirname, './variables'),
-      'variables/icons': path.resolve(__dirname, './variables/icons'),
+      constants: path.resolve(__dirname, './constants'),
       assets: path.resolve(__dirname, './assets'),
       tests: path.resolve(__dirname, './tests'),
       context: path.resolve(__dirname, './context'),
